Exceeções e Correções

+++ src/main/java/br/org/unicortes/barbearia/repositories/SaleRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface SaleRepository extends JpaRepository<Sale, Long> {
 
-    Sale findBySaleId(int id);
+    Sale findBySaleId(Long id);
 }
